feat(#5490): merge explorer into analysis board
